+++ platformio/__init__.py
@@ -12,20 +12,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-VERSION = (6, 1, 9)
+VERSION = (6, 1, 10)
 __version__ = ".".join([str(s) for s in VERSION])
 
 __title__ = "platformio"
 __description__ = (
-    "A professional collaborative platform for embedded development. "
-    "Cross-platform IDE and Unified Debugger. "
-    "Static Code Analyzer and Remote Unit Testing. "
-    "Multi-platform and Multi-architecture Build System. "
-    "Firmware File Explorer and Memory Inspection. "
-    "IoT, Arduino, CMSIS, ESP-IDF, FreeRTOS, libOpenCM3, mbedOS, Pulp OS, SPL, "
-    "STM32Cube, Zephyr RTOS, ARM, AVR, Espressif (ESP8266/ESP32), FPGA, "
-    "MCS-51 (8051), MSP430, Nordic (nRF51/nRF52), NXP i.MX RT, PIC32, RISC-V, "
-    "STMicroelectronics (STM8/STM32), Teensy"
+    "Your Gateway to Embedded Software Development Excellence. "
+    "Unlock the true potential of embedded software development "
+    "with PlatformIO's collaborative ecosystem, embracing "
+    "declarative principles, test-driven methodologies, and "
+    "modern toolchains for unrivaled success."
 )
 __url__ = "https://platformio.org"
 
@@ -46,7 +42,7 @@ __core_packages__ = {
     "contrib-piohome": "~3.4.2",
     "contrib-pioremote": "~1.0.0",
     "tool-scons": "~4.40502.0",
-    "tool-cppcheck": "~1.270.0",
+    "tool-cppcheck": "~1.21100.0",
     "tool-clangtidy": "~1.150005.0",
     "tool-pvs-studio": "~7.18.0",
 }
@@ -56,3 +52,22 @@ __check_internet_hosts__ = [
     "88.198.170.159",  # platformio.org
     "github.com",
 ] + __registry_mirror_hosts__
+
+__install_requires__ = [
+    # Core requirements
+    "bottle == 0.12.*",
+    "click >=8.0.4, <=8.2",
+    "colorama",
+    "marshmallow == 3.*",
+    "pyelftools == 0.29",
+    "pyserial == 3.5.*",  # keep in sync "device/monitor/terminal.py"
+    "requests == 2.*",
+    "semantic_version == 2.10.*",
+    "tabulate == 0.*",
+] + [
+    # PIO Home requirements
+    "ajsonrpc == 1.2.*",
+    "starlette >=0.19, <0.32",
+    "uvicorn >=0.16, <0.24",
+    "wsproto == 1.*",
+]